feat: move types to separate file
