+++ src/configTree.tsx
@@ -1,10 +1,4 @@
-export type ID = string | number
-
-export interface Node {
-  id: ID
-  name: string
-  children?: Node[]
-}
+import { Node } from './global.types'
 
 export const configTree: Node[] = [
   {
