+++ src/global.types.ts
@@ -1,0 +1,7 @@
+export type ID = string | number
+
+export interface Node {
+  id: ID
+  name: string
+  children?: Node[]
+}
